Cleanup a few scheduler warts.
Functions are now first class targets for a plan's execution.  The
example thrift planners are converted from tasks to functions to
exercise this.

Scheduler is now a proper `AbstractClass`.  A bad implementatiion of
`Plan.__getattr__` is fixed to allow this.  The same bad implementation
of `__getattr__` is also fixed in `Configuration`.