Add missing docs and stablize scheduler test.
The scheduler test was not stable due to aggregation by the
GlobalIvyResolvePlanner over subjects.  Trim back the test to be as it
was and just assert product_type, plan pairs since the particular
subject of the exemplar promise paired to the plan is not important
(plans are registered against all their subjects by
`ProductMapper.register_promises`).