Add fail-slow handling to `Engine`.
This allows for an engine execution to be in `fail_slow` mode, in which
case as much of the execution graph as can be executed is and failure
paths are recorded using a `FailedToProduce` product type.

A test is added with a failing path and a good path to exercise this
handling in both the serial and multiprocess engines.

Testing Done:
CI went green here:
  https://travis-ci.org/pantsbuild/pants/builds/87523475

Bugs closed: 915, 2452, 2458

Reviewed at https://rbcommons.com/s/twitter/r/3040/