Defer argparse registration to the last possible moment.
Now option registration just records the registration args,
and they get applied as needed, as an implementation detail.

This was also an opportunity to clean up and de-complicate
how we register recursive options, so the code is neater now.

The performance benefits are very noticeable in profiles.

Note that this means that some validation will only happen if
the options are actually requested, which required a few
changes in tests.