Better implementation of the reporting emitter thread.
Previously it used twitter.commons.threading.PeriodicThread, which
it turns out just calls time.sleep() in a loop. So we were wasting
up to half a second on a no-op pants shutdown just for that thread
to exit.

Now we use proper event signaling, so the thread can be interrrupted
immediately.

This also gets rid of the t.c.threading dep entirely.