Improve incremental caching tests
While attempting to capture the failure I was seeing yesterday, I made `test_task.py` way more complicated than it needs to be. But the reason I was having trouble reproducing it was that it was exactly the existing ABA case that Nick asked for a test for. But because the test wasn't inspecting the `previous_results_dir`, it wasn't noticing that it wasn't being set.

- Revert test changes from [r/3028](https://rbcommons.com/s/twitter/r/3028/)
- Test the existence of the `previous_results_dir` instead

Testing Done:
https://travis-ci.org/pantsbuild/pants/builds/87216048

Bugs closed: 2455

Reviewed at https://rbcommons.com/s/twitter/r/3034/