Change JarPublish.fingerprint to JarPublish.entry_fingerprint to ensure task fingerprint can change
While doing some debugging in intellij, I noticed that JarPublish's fingerprint method was hilighted with a warning that it overrode a parent with a different signature. This lead me to realize that its fingerprint is now going to be a function rather than a string containing a hash. That didn't seem very good.

This renames the JarPublish fingerprint method so it is no longer overriding the property

Testing Done:
Saw that it doesn't override Task.fingerprint. CI away on the PR.

Bugs closed: 2502

closes #2502

Reviewed at https://rbcommons.com/s/twitter/r/3078/