Improve `_gather_one_pending` docs.
Made the motivation for this method more clear relative to the most
obvious way to get back results from the pool as they complete.

Also Make the `_panding_results` list type uniformly `AsyncResult`.

The `AsyncResult` type, although part of the multiprocessing package
public docs was not exposed in all environments.  As such, kill the type
test in `_gather_one_pending` by creating a pre-satisfied `AsyncResult`
duck-type for the `FailedToProduce` case to use.