Gettings the ./pants junit runner to work with Cucumber scenarios.
Cucumber Scenarios generate test Descriptions that are unusually
formatted. Cucumber sends the Scenario description as the
"class name", and the step description as the "method name". The
actual test class object is null.

This was causing the junit runner itself to generate NPEs and fail
on some of our internal targets which use Cucumber.

This fixes the problem by making the AntJunitXmlReportListener less
dependent on the specifics of what's stored in the class name and
method name fields.