DO NOT MERGE: Demo a bug where cucumber breaks the junit runner.
Cucumber Scenarios generate test Descriptions that are unusually
formatted. Cucumber sends the Scenario description as the
"class name", and the step description as the "method name". The
actual test class object is null.

This causes the junit runner itself to generate NPEs and fail on
some of our internal targets which use Cucumber. This branch
demonstrates the failure.