Add `pants_requirement()` for plugin authors.
When writing a plugin outside the `pantsbuild/pants` repo, if you want
to use pants itself to build the plugin or test it, you need to express
a BUILD dependency on pants itself.  This should always match the
running pants and for users of the `pantsbuild.pants` official sdists,
this is awkward.  The `pants_requirement()` macro provides a clean way
to express the plugin's dependency on the correct pants distribution.