Extract `python-eval` and `pythonstyle` to plugin.
This moves these two tasks to a python checks plugin in contrib and adds
registration in the 'compile' goal to enable easy install and activation
of these extra checks using the `plugins` mechanism; ie a pure
`pants.ini` activation.

To enable use of individual checks but not all, a `--skip` is added to
`python-eval`.