Allow bundle to run for all targets, rather than just target roots
Problem

We are developing an integration test framework where we will start JVM applications from within a JVM test. As part of this, we need to build the bundles for dependency applications and to have all this happen automatically, we want to specify the jvm_app targets as dependencies for our junit_tests target.

Unfortunately, this currently doesn't work as the bundle task only runs for the target roots, not all targets so calling bundle on the junit_tests target will not call bundle for its dependencies.

Solution

Change the bundle_create task to run for all targets, not just target roots.

This is in general safe as jvm_app targets should not have dependees themselves and targets should in general depend on library targets, not on jvm_app (or jvm_binary) targets.

Bugs closed: 2558

Reviewed at https://rbcommons.com/s/twitter/r/3119/