Deprecate calling with_description() when registering a task.
The problem is that this description was associated with the
*goal* the task was registered on, not the task itself.

This meant that we were dependent on registration order. More
problematically, it meant that, e.g., the JVM backend was
creating descriptions for generic goals such as compile and test.

Instead, we introduce an explicit Goal.register(name, description)
solely for providing descriptions for generic goals.

Then we enumerate the generic goals in core_tasks/register.py.

Single-task goals take their description from the task's docstring,
so this change also ensures that we have appropriate docstrings
everywhere.