Refactoring to allow NodeModule subclassing
This is a large commit that does two primary things:

Refactors NPM-named things to Node
Taking advantage of that refactoring, introduces a node_preinstalled_module target type that downloads a prepackaged node_modules directory in a tar.gz file from a URL. This is somewhat useful, but it's more meant as an example to show subclassing NodeModule and using a different resolver.
These address the problems outlined under #2642.

Right now, the example used in the integration test for has a URL pointing to my own GitHub. I need to add that file to Bintray somehow, but I need @jsirois's help for that.

Testing Done:
https://travis-ci.org/pantsbuild/pants/builds/95002417

Bugs closed: 2641

Reviewed at https://rbcommons.com/s/twitter/r/3192/