Move resources() and prep_command() out of backend/core.
doc.py is the last target type in backend/core. I'll move that
in a future change, but to accommodate it now I created a new
target just for it, to isolate deps on it from deps on the
other stuff in backend/core/targets.

Also created a new target for dependencies.py, so we can remove
the remaining internal deps on backend/core/targets:all/common.