prefer explicit jvm locations over internal heuristics
Previously if --jvm-distributions-paths was not provided, pants
searched for jvms in the folling order:
 (A) JAVA_HOME/JDK_HOME
 (B) Look at common file system locations
 (C) java on PATH

This commit change that order to always prefer the explicit user
configurations (A, C) over heuristics (B).  Besides just being more
consistent, this means when a user gets an error about java not
working when invoked by pants, they are more likely to be using the
same jdk when they try to reproduce it.

An additional benefit is that the user chosen jvm likely "works"
(without pants needing to maintain careful min/max version constraints
for every too), while old EOL things in /usr/lib/jvm may not.  For
example our Centos6 CI server has the super old `java-1.5.0-gcj`
package, which does not work with modern ivy.

fixes #2682