Prepend '//' to Address.spec when the spec_path is empty.
Also fix a bug in Address.reference that this revealed: namely
that spec_path-relative references at the root were actually
going through the code path for absolute references, but
because the result was identical it didn't matter.

An issue in https://rbcommons.com/s/twitter/r/3225/ uncovered a bug in the way we
handle addresses at the root of the repo (i.e. with an empty spec_path) in the
context of Target.traversable_dependency_specs.  These specs are always with respect
to the target being loaded, meaning that relative spec paths are relative to
the target's spec_path.  This means that specs with an empty spec_path must
be disambiguated by prepending //, otherwise they will be treated as a relative
spec.

Testing Done:
Had to fix up lots of little assertions that rely on the spec formatting.

CI is green: https://travis-ci.org/pantsbuild/pants/builds/96364177

Bugs closed: 2693

Reviewed at https://rbcommons.com/s/twitter/r/3234/