Fixup the `NodePreinstalledModuleResolver`.
Previously, it assumed the `/tmp` extraction dir was on the same
filesystem as the build root (it used `os.rename`).  This change fixes
the rename to use `shutil.move` which also uses `os.rename` - but only
when appropriate.

The existing `NodeResolveIntegrationTest`
`test_resolve_preinstalled_node_module_project` failed on machines with
`/tmp` on a seperate filesystem from the build root and now passes on
those machines.