Make it easy for tests to use the "real" python interpreter cache.
This greatly speeds up tests that need to resolve an interpreter.
We already had this ability in PythonTaskTestBase, I just moved it
to a InterpreterCacheTestMixin, for easier re-use.

I then applied it to test_export.py, which was a very slow set of
tests (around a minute of runtime) due to repeated resolution of
interpreters. With this change those tests run in a 3-5 seconds.