A graph_info backend.
Contains all the tasks that perform computation
on the build graph, and that used to be in backend/core.

While in that code, rename and tag an integration test
that was masquerading as a unittest and taking about 25%
of the total unit test exec time.

With this change, the elimination of backend/core is almost
complete!

Testing Done:
CI passes: https://travis-ci.org/pantsbuild/pants/builds/96881334

Reviewed at https://rbcommons.com/s/twitter/r/3256/