Fixup `GoFetch` to respect transitive injections.
Previously, a fetch could trigger injection of a `GoRemoteLibrary`
subgraph that was not tracked by `GoFetch` leading to the erroneous
assumption that it was safe to synthesize a target that was already
implicitly injected into the active graph.

Add a failing test for this scenario that is fixed by this change.