Fixup `GoFetch` to respect transitive injections.
Previously, a fetch could trigger injection of a `GoRemoteLibrary`
subgraph that was not tracked by `GoFetch` leading to the erroneous
assumption that it was safe to synthesize a target that was already
implicitly injected into the active graph.

Add a failing test for this scenario that is fixed by this change.

Testing Done:
I tested this in the Medium/mono repo and the change does fix their
problem.  A new unit test also emulates the problem and is likewise
fixed.

CI went green here:
  https://travis-ci.org/pantsbuild/pants/builds/97707029

Bugs closed: 2733, 2740

Reviewed at https://rbcommons.com/s/twitter/r/3270/