Simplify build dict/reference information extraction.
Previously we had various non-standard ways of providing
BUILD file objects and context-aware object factories: Some
were objects with __call__ methods, some were functions, some
were the results of currying the context arg.

All this made it difficult to introspect on these to identify
the args and docstrings used to generate the build dictionary.

This change standardizes how these objects are provided, and makes
corresponding changes to the info extracters to take advantage
of this.

A future change will use this to rewrite the build dictionary/reference
page generator, so we can get rid of reflect.py.