Multiple dependency_managements with multiple ivy resolves.
This is the second pass of implementing dependency management in
pants.

The first pass was reviewed here:

    https://rbcommons.com/s/twitter/r/3336/

The design document is available here:

    https://docs.google.com/document/d/1AM_0e1Az_NHtR150Zsuyaa6u7InzBQGLq8MrUT57Od8/edit#

This change allows managed_jar_dependencies targets to be chained
together; the set of pinned artifact is the union of all jar
coordinates specified by the transitive closure of the target
specified in a jar_library's managed_dependencies field.

When there are conflicts between dependencies in the transitive
closure, an error is raised.

The change to make ivy do multiple resolves is fairly simple; see
ivy_task_mixin.py's resolve() method.