A managed_jar_libraries factory to reduce 3rdparty duplication.
In an effort to reduce duplication of 3rdparty configuration, I
introduced a `managed_jar_libraries` factory. It syntactically
appears very similarly to a managed_jar_dependencies target, but it
also generates jar_library targets for each jar() it contains. It
does not generate anything for artifacts which are specified
indirectly with a jar_library spec (rather than a directly embedded
jar() object), because that would be redundant (and could even
likely conflict with existing definitions in the BUILD file).