Separate the ivy resolution and ivy repository cache directories.
Currently, the ivy cache directories are unified and by default
live in a shared folder under the user's home directory (~/.ivy2/pants).
While this is fine for fetched artifacts (there is a lock strategy),
it is explicitly documented by Ivy to be unsafe for concurrent
access (http://ant.apache.org/ivy/history/2.4.0/concept.html#cache).
Since multiple processes running pants in separate workspaces
should not interfere with one another, the resolution cache should
therefore be split out.

One option is to always set the default ivy cache (both resolution
and repository) to a workspace-local directory.  This is certainly
safe, but it is not performant since a clean of the workspace
(or clean-all if the directory is under the workdir) will result in
all downloaded artifacts being deleted.  This is probably okay for
small projects, but not projects with large numbers of transitively
resolved artifacts.

This change favors backwards compatibility by leaving the unsafe
strategy as the default, but adds a new option to control where
the resolution dir lives.  With a properly set up ivysettings.xml,
this gives us the best of both worlds (correctness vs performance)
and paves the way for always separating these two directories.

The caveat is that this resolution dir must be known to pants, since
pants must extract and parse the resolution reports after invoking
Ivy.  Thus, if pants plumbs this directory, Ivy _must_ use it or
the resolution post-processing will fail.  Unfortunately, the
setting cannot be plumbed directly.  Instead we pass in a JVM
property which the user is required to consume properly in their
ivysettings.xml.

By default, an ivysettings.xml is not required.  Therefore we make sure
to error out if a custom resolution directory is set but a custom
ivysettings.xml is not.  Because of the flexible nature of ivy
configurations, it is not feasible to simply parse and lint against
the use of the property (ivy has "includes" and the provided settings.xml
might just be a thin wrapper over others).

In the future we should capture the relevant structures that ivy
settings can express and generate the ivysettings.xml ourselves always.
Alternatively, splitting resolution and fetching in a way that fully
captures both resolution and fetching in the pants artifact cache
might make the (currently unperformant) strategy of always keeping
resolution and repository directories within the current workspace
the more desirable option.