Don't skip writing <artifact>s to ivy.xml even if there's only one.
We ran into a problem internally at Square where ivy would
stubbornly resolve a 'test-jar' instead of a 'jar' artifact, even
when we directly declared a dependency on the 'jar'. The reason for
this appears to be that pants was short-circuiting instead of
writing an explicit artifact for the 'jar'.

This changes:

   <dependency org="org.apache.hadoop"
                  name="hadoop-common"
                  rev="2.0.0-cdh4.6.0">
        <conf name="default" mapped="default"/>
        <!-- excludes, etc -->
    </dependency>

To:

   <dependency org="org.apache.hadoop"
                  name="hadoop-common"
                  rev="2.0.0-cdh4.6.0">
        <conf name="default" mapped="default"/>
        <artifact
          name="hadoop-common"
          type="jar"/>
        <!-- excludes, etc -->
    </dependency>

Which causes ivy to behave as expected.