Adding support for multiline param help descriptions in Pants BUILD Dictionary
In response to a user reported error that target parameter help appeared incomplete, added support for parameter help strings that span multiple lines.

Prior to this change, get_arg_descriptions_from_docstring did not suppport multiline descriptions, as was noted in the comments.  However, there are numerous instances in which existing parameter help spans multiple lines, and there's nothing enforcing the single-line description convention.  In addition, especially for parameters that have longer names, it may be difficult to include meaningful descriptions that span only a single line.  Thus, it seemed best to go ahead and add support for multiline descriptions.

Adding support for multiline descriptions also fixed the fact that no help was appearing in the Pants BUILD Dictionary for parameters where the help string included :param name: on one line and the description on the next line.   This can be seen in the help for the provides parameter of contrib_plugin (screenshots attached).

As seen in the screenshots for android_binary below, adding support for mulitline descriptions altered the formatting of the tables containing those descriptions.  In particular, the parameter names sometimes wrap, and the description text is aligned in the middle of the cell.  If these changes go in, I'll open another github issue on that, given my very limited css skills.

Testing Done:
https://travis-ci.org/pantsbuild/pants/builds/106130379

Bugs closed: 2848

Reviewed at https://rbcommons.com/s/twitter/r/3399/