If the junit output file doesn't exist, it should still count as an error on the target
We have some junit tests in our repo that fail outside of an actual testmethod, so the junit xml output file doesn't get created. If that file doesn't get created, it's a failure of the target, so that target should still be in TestFailedTaskError even though we don't have a testmethod and testclass.

Testing Done:
build-support/bin/unit-tests.sh

The tests that fail in our repo are quite inscrutable so I can't tell how to create a simple unit test for this. Would it be ok to submit this as-is?

It works on the degenerate tests in our repo.

Reviewed at https://rbcommons.com/s/twitter/r/3407/