Retain file permissions when shading monolithic jars.
On some systems (macos at least), the atomic_copy function at the
end of the jar shading process would cause the file permissions to
be set to '-rw-------', instead of the normal '-rw-r--r--', because
of the permissions of the temporary file that atomic_copy creates.

This change simply records the permissions of the jar before
shading, and chmod's the shaded jar to have whatever permissions it
had previously.