add public api markers
Based on a discussion on pants slack, a method of marking methods as public using docstrings was proposed.  This RB is an
example of how this markup would happen.  This is part of the effort discussed in:
https://docs.google.com/document/d/1HDIalsQRv2Ds_uylPznX_VPmMdJib-1PdaGNVvF7fu4/edit#

* Add markers to public api methods and classes

Testing Done:
passes on personal fork.
CI green: https://travis-ci.org/pantsbuild/pants/builds/106836984

Reviewed at https://rbcommons.com/s/twitter/r/3417/