support searching multiple linux java dist dirs
SUSE is an odd ball out and does not use `/usr/lib/jvm`.  This makes
the unit tests somewhat more awkward with the temporary dirs and
reaching into the class.  I think cleaner long term approach would be
to use a fake file sytem (such as pyfakefs) instead of actual temp
dirs, but that is beyond the scope of this change.

fixes #2931