Change the mechanism for detecting the buildroot.
The new mechanism uses the cwd, but heuristically verifies
that there's a pants runner script in it.

The old mechanism climbed up the directory hierarchy until it
found a pants.ini file.

This change is because:

A) We don't want pants.ini to carry two unrelated meanings (config file,
   and build root indicator).
B) We want to support having the config file anywhere, and under any name,
   as determined by an option (a pending change implements this).
C) We want Pants to be able to run with no pants.ini at all, for
   conceptual neatness. "You must have at least an empty config file" is an
   odd requirement.
D) The directory climbing was overkill in practice, because running
   pants from anywhere other than the buildroot doesn't work anyway
   for a variety of reasons.

Note that this change doesn't preclude supporting running pants from
a subdir in the future, but for now it fails fast on this case,
instead of failing in some confusing way later on.