Change how we detect the buildroot.
Instead of looking for pants.ini we look for the pants runner script
(or any file named 'pants') in the cwd and its ancestors.

This is because we don't want to give the location of pants.ini
any special extra meaning.  We want pants to work with multiple cascading
config files (instead of the clunky --config-file-override mechanism)
or even no config file at all. A future change will implement configurable
config file locations, and will explain in greater detail why that's a
good thing...