reset Subsystems when creating a new context in tests
Subsystem instances may cache options values, so it's not sufficient to just assign to Subsystem._options, we also need to clear out the subsystem instances.

This change causes context creation to clear the global subsystems, and fixes the one test that had issues with doing so to use options rather than reaching into the subsystem instance directly.