Change ivy resolve ordering to attempt load first and fall back to full resolve if load fails.
I've been working on splitting resolve into a resolve and fetch. This patch backports the structural approach I've ended up with. In it, we bias towards the faster, loading paths first, checking for their preconditions and falling through to the calling out to ivy paths if they aren't met.

The current implementation instead checks to see if work is needed, then falls through to loading if it is not. The problem with that approach is that if it turns out that the loaded resolve is incomplete, we can't complete it--we'll blow up somewhere later. This problem would become bigger with more potential loading and resolving paths. Moving to the new model makes adding more loading and resolving mechanisms easier.

- Move ivy reports to resolve workdir after successful resolves.
- Introduce class to encapsulate the result of a ivy resolve.
- Fall back to a full resolve when any ivy symlink points to a missing file.
- Bump the implementation version.
- Make parsing an Ivy report via a path public.
- Update docstrings for resolve and ivy_classpath.

Testing Done:
Running jvm tests in a large subgraph locally. CI passed at https://travis-ci.org/pantsbuild/pants/builds/112160887

Bugs closed: 2967

Reviewed at https://rbcommons.com/s/twitter/r/3501/

closes #2967